Finish configuration of GitHub actions runner.
The playbook can be executed multiple times to ensure the
service is running. Currently reconfiguration of the node
requires the machine to be torn down and replaced.
This is not necessarily a bad thing but is not yet automated.

Also the removal of the actions runner from the GitHub runners
list must be done manually currently.